FUSETOOLS2-1168 - snippet for Jbang task
