+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 ## 0.2.0
 
 - Provide error message and remind requirements in case of failure when trying to attach the debugger
-- Provide configuration snippets for Camel debugger launch configuration and Camel application launch task
+- Provide configuration snippets for Camel debugger launch configuration
+- Provide configuration snippets to launch Camel application ready to accept a Camel debugger connection using jbang or Maven
 
 ## 0.1.1
